+++ common/src/main/java/tocraft/remorphed/handler/LivingDeathHandler.java
@@ -9,6 +9,7 @@ import tocraft.craftedcore.event.common.EntityEvents;
 import tocraft.craftedcore.patched.CEntity;
 import tocraft.remorphed.Remorphed;
 import tocraft.remorphed.impl.PlayerMorph;
+import tocraft.walkers.Walkers;
 import tocraft.walkers.api.PlayerShape;
 import tocraft.walkers.api.PlayerShapeChanger;
 import tocraft.walkers.api.variant.ShapeType;
@@ -18,7 +19,7 @@ public class LivingDeathHandler implements EntityEvents.LivingDeath {
     public InteractionResult die(LivingEntity entity, DamageSource source) {
         if (!(entity instanceof Player) && source.getEntity() instanceof ServerPlayer killer) {
             ShapeType<?> type = ShapeType.from(entity);
-            if (type != null) {
+            if (type != null && (!Walkers.CONFIG.blacklistPreventsUnlocking || !Walkers.isPlayerBlacklisted(killer.getUUID()))) {
                 PlayerMorph.addKill(killer, type);
 
                 if (Remorphed.CONFIG.autoTransform && PlayerMorph.getKills(killer, type) >= Remorphed.getKillToUnlock(type.getEntityType())) {